modified:   src/testing.py 	modified:   src/__pycache__/discord_stuff.cpython-310.pyc 	modified:   src/__pycache__/generate_tweet.cpython-310.pyc 	modified:   src/__pycache__/tweets_class.cpython-310.pyc 	modified:   src/data/tweets.csv 	modified:   src/discord_stuff.py 	modified:   src/generate_tweet.py 	modified:   src/main.py 	modified:   src/tweets_class.py
Added a function to check sentiment in the generate_tweet file. This function is called in the main file and the sentiment is added to the tweet. The sentiment is also added to the csv file.
Amended the main.py to call a function for retrieving tweets, hopefully get this running every hour.